Module twilight_http::request::application

Structs

CreateFollowupMessage
Create a followup message to an interaction.

CreateGlobalCommand
Create a new global command.

CreateGuildCommand
Create a new command in a guild.

DeleteFollowupMessage
Delete a followup message created from a interaction.

DeleteGlobalCommand
Delete a global command, by ID.

DeleteGuildCommand
Delete a command in a guild, by ID.

DeleteOriginalResponse
Delete a original interaction response.

GetCommandPermissions
Fetch command permissions for a command from the current application in a guild.

GetGlobalCommands
Retrieve all global commands for an application.

GetGuildCommandPermissions
Get command permissions for all commands from the current application in a guild.

GetGuildCommands
Fetch all commands for a guild, by ID.

InteractionCallback
Respond to an interaction, by ID and token.

InteractionError
The error created if the creation of interaction fails.

SetCommandPermissions
Update command permissions for all commands in a guild.

SetGlobalCommands
Set global commands.

SetGuildCommands
Set a guild’s commands.

UpdateCommandPermissions
Update command permissions for a single command in a guild.

UpdateFollowupMessage
Update a followup message.

UpdateFollowupMessageError
A followup message can not be updated as configured.

UpdateGlobalCommand
Edit a global command, by ID.

UpdateGuildCommand
Edit a command in a guild, by ID.

UpdateOriginalResponse
Update the original response created by a interaction.

UpdateOriginalResponseError
A original response can not be updated as configured.

Enums

InteractionErrorType

UpdateFollowupMessageErrorType
Type of UpdateFollowupMessageError that occurred.

UpdateOriginalResponseErrorType
Type of UpdateOriginalResponseError that occurred.