Module twilight_http::request::application::interaction

Structs

CreateFollowup

Create a followup message to an interaction, by its token.

CreateResponse

Respond to an interaction, by its ID and token.

DeleteFollowup

Delete the original message, by its token.

DeleteResponse

Delete a followup message to an interaction, by its token and message ID.

GetFollowup

Get a followup message of an interaction, by its token and the message ID.

GetResponse

Get the original message, by its token.

UpdateFollowup

Edit a followup message of an interaction, by its token and the message ID.

UpdateResponse

Edit the original message, by its token.