Struct twilight_http::request::channel::message::allowed_mentions::AllowedMentionsBuilder

pub struct AllowedMentionsBuilder<'a, E, U, R> { /* fields omitted */ }

A builder for allowed mentions.

Example

use twilight_http::request::channel::message::allowed_mentions::AllowedMentionsBuilder;
let mut allowed_mentions = AllowedMentionsBuilder::new()
    .parse_everyone()
    .parse_users()
    .build_solo();

Implementations

impl<'a> AllowedMentionsBuilder<'a, Unspecified, Unspecified, Unspecified>

pub fn new() -> Self

Create the builder.

impl<'a, U, R> AllowedMentionsBuilder<'a, Unspecified, U, R>

pub fn parse_everyone(self) -> AllowedMentionsBuilder<'a, Parsed, U, R>

Enable parsing for the @everyone and @here tags.

impl<'a, E, R> AllowedMentionsBuilder<'a, E, Unspecified, R>

pub fn parse_users(self) -> AllowedMentionsBuilder<'a, E, Parsed, R>

Enable parsing for all user tags.

pub fn parse_specific_users(
self,
u: impl IntoIterator<Item = UserId>
) -> AllowedMentionsBuilder<'a, E, ExplicitUser, R>

Enable parsing for specific user tags.

impl<'a, E, U> AllowedMentionsBuilder<'a, E, U, Unspecified>

pub fn parse_roles(self) -> AllowedMentionsBuilder<'a, E, U, Parsed>

Enable parsing for all role tags.

pub fn parse_specific_roles(
self,
r: impl IntoIterator<Item = RoleId>
) -> AllowedMentionsBuilder<'a, E, U, ExplicitRole>

Enable parsing for specific role tags.

impl<'a, E, U> AllowedMentionsBuilder<'a, E, U, ExplicitRole>

pub fn parse_specific_roles(self, r: impl IntoIterator<Item = RoleId>) -> Self

Enable parsing for more specific role tags.

If there are already some specific RoleIds in this builder, extend them with the content of r.

impl<'a, E, R> AllowedMentionsBuilder<'a, E, ExplicitUser, R>

pub fn parse_specific_users(self, u: impl IntoIterator<Item = UserId>) -> Self

Enable parsing for more specific user tags.

If there are already some specific UserIds in this builder, extend them with the content of u.

impl<'a, E: VisitAllowedMentionsEveryone, U: VisitAllowedMentionsUsers, R: VisitAllowedMentionsRoles> AllowedMentionsBuilder<'a, E, U, R>

pub fn build(self) -> CreateMessage<'a>ⓘ

Important traits for CreateMessage<'_>
impl<'_> Future for CreateMessage<'_> type Output = Result<Message>;

Return a CreateMessage struct with the specified allowed_mentions.

pub fn build_solo(self) -> AllowedMentions

Build a raw AllowedMentions for use in ClientBuilder#default_allowed_mentions.