Struct twilight_http::request::channel::message::update_message::UpdateMessage

pub struct UpdateMessage<'a> { /* fields omitted */ }

Update a message by ChannelId and MessageId.

You can pass None to any of the methods to remove the associated field. For example, if you have a message with an embed you want to remove, you can use .embed(None) to remove the embed.

Examples

Replace the content with "test update":

use twilight_http::Client;
use twilight_model::id::{ChannelId, MessageId};

let client = Client::new("my token".to_owned());
client.update_message(ChannelId(1), MessageId(2))
    .content(Some("test update"))?
    .exec()
    .await?;

Remove the message’s content:

client.update_message(ChannelId(1), MessageId(2))
    .content(None)?
    .exec()
    .await?;

Implementations

impl<'a> UpdateMessage<'a>

pub const fn attachments(self, attachments: &'a [Attachment]) -> Self

Specify multiple attachments already present in the target message to keep.

If called, all unspecified attachments will be removed from the message. If not called, all attachments will be kept.

Calling this method will clear any previous calls.

pub fn components(
self,
components: Option<&'a [Component]>
) -> Result<Self, UpdateMessageError>

Add multiple Components to a message.

Calling this method multiple times will clear previous calls.

Pass None to clear existing components.

Errors

Returns an UpdateMessageErrorType::ComponentCount error type if too many components are provided.

Returns an UpdateMessageErrorType::ComponentInvalid error type if one of the provided components is invalid.

pub fn content(
self,
content: Option<&'a str>
) -> Result<Self, UpdateMessageError>

Set the content of the message.

Pass None if you want to remove the message content.

Note that if there is no embed then you will not be able to remove the content of the message.

The maximum length is 2000 UTF-16 characters.

Errors

Returns an UpdateMessageErrorType::ContentInvalid error type if the content length is too long.

pub fn embeds(self, embeds: &'a [Embed]) -> Result<Self, UpdateMessageError>

Attach multiple embeds to the message.

To keep all embeds, do not use this.

To modify one or more embeds in the message, acquire them from the previous message, mutate them in place, then pass that list to this method.

To remove all embeds pass an empty slice.

Note that if there is no content or file then you will not be able to remove all of the embeds.

Calling this method will clear any previous calls.

Errors

Returns an UpdateMessageErrorType::EmbedTooLarge error type if an embed is too large.

pub const fn suppress_embeds(self, suppress: bool) -> Self

Suppress the embeds in the message.

pub fn allowed_mentions(self, allowed: AllowedMentions) -> Self

Set the AllowedMentions in the message.

pub fn exec(self) -> ResponseFuture<Message>

Execute the request, returning a future resolving to a Response.