Struct twilight_http::request::channel::UpdateChannel

pub struct UpdateChannel<'a> { /* private fields */ }

Update a channel.

All fields are optional. The minimum length of the name is 1 UTF-16 character and the maximum is 100 UTF-16 characters.

Implementations

impl<'a> UpdateChannel<'a>

pub const fn bitrate(self, bitrate: u32) -> Self

Set the bitrate of the channel. Applicable to voice channels only.

pub fn name(self, name: &'a str) -> Result<Self, ChannelValidationError>

Set the name.

The minimum length is 1 UTF-16 character and the maximum is 100 UTF-16 characters.

Returns an error of type NameInvalid if the name is invalid.

pub const fn nsfw(self, nsfw: bool) -> Self

Set whether the channel is marked as NSFW.

pub const fn parent_id(self, parent_id: Option<Id<ChannelMarker>>) -> Self

If this is specified, and the parent ID is a ChannelType::CategoryChannel, move this channel to a child of the category channel.

pub const fn permission_overwrites(
self,
permission_overwrites: &'a [PermissionOverwrite]
) -> Self

Set the permission overwrites of a channel. This will overwrite all permissions that the channel currently has, so use with caution!

pub const fn position(self, position: u64) -> Self

Set the position of the channel.

Positions are numerical and zero-indexed. If you place a channel at position 2, channels 2-n will shift down one position and the initial channel will take its place.

pub const fn rate_limit_per_user(
self,
rate_limit_per_user: u16
) -> Result<Self, ChannelValidationError>

Set the number of seconds that a user must wait before before they are able to send another message.

The minimum is 0 and the maximum is 21600. This is also known as “Slow Mode”. See Discord Docs/Channel Object.

Errors

Returns an error of type RateLimitPerUserInvalid if the name is invalid.

pub fn topic(self, topic: &'a str) -> Result<Self, ChannelValidationError>

Set the topic.

The maximum length is 1024 UTF-16 characters. See Discord Docs/Channel Object.

Errors

Returns an error of type TopicInvalid if the name is invalid.

pub const fn user_limit(self, user_limit: u16) -> Self

For voice channels, set the user limit.

Set to 0 for no limit. Limit can otherwise be between 1 and 99 inclusive. See Discord Docs/Modify Channel.

pub const fn video_quality_mode(
self,
video_quality_mode: VideoQualityMode
) -> Self

Set the VideoQualityMode for the voice channel.

pub const fn kind(self, kind: ChannelType) -> Self

Set the kind of channel.

Only conversion between ChannelType::GuildText and ChannelType::GuildNews is possible, and only if the guild has the NEWS feature enabled. See Discord Docs/Modify Channel.

pub fn exec(self) -> ResponseFuture<Channel>

Execute the request, returning a future resolving to a Response.