Module twilight_http::request::channel::thread

Re-exports

pub use self::create_forum_thread::CreateForumThread;

Modules

create_forum_thread

Structs

AddThreadMember

Add another member to a thread.

CreateThread

Start a thread that is not connected to a message.

CreateThreadFromMessage

Create a new thread from an existing message.

GetJoinedPrivateArchivedThreads

Returns archived private threads in the channel that the current user has joined.

GetPrivateArchivedThreads

Returns archived private threads in the channel.

GetPublicArchivedThreads

Returns archived public threads in the channel.

GetThreadMember

Returns a ThreadMember in a thread.

GetThreadMembers

Returns the ThreadMembers of the thread.

JoinThread

Add the current user to a thread.

LeaveThread

Remove the current user from a thread.

RemoveThreadMember

Remove another member from a thread.

UpdateThread

Update a thread.