Struct twilight_http::request::channel::thread::create_forum_thread::CreateForumThreadMessage

pub struct CreateForumThreadMessage<'a>(/* private fields */);

Implementations

impl<'a> CreateForumThreadMessage<'a>

pub fn allowed_mentions( self, allowed_mentions: Option<&'a AllowedMentions>, ) -> Self

Specify the AllowedMentions for the message.

Unless otherwise called, the request will use the client’s default allowed mentions. Set to None to ignore this default.

pub fn attachments(self, attachments: &'a [Attachment]) -> Self

Attach multiple files to the message.

Calling this method will clear any previous calls.

Errors

Returns an error of type AttachmentFilename if any filename is invalid.

pub fn components(self, components: &'a [Component]) -> Self

Set the message’s list of Components.

Calling this method will clear previous calls.

Requires a webhook owned by the application.

Errors

Refer to the errors section of twilight_validate::component::component for a list of errors that may be returned as a result of validating each provided component.

pub fn content(self, content: &'a str) -> Self

Set the message’s content.

The maximum length is 2000 UTF-16 characters.

Errors

Returns an error of type ContentInvalid if the content length is too long.

pub fn embeds(self, embeds: &'a [Embed]) -> Self

Set the message’s list of embeds.

Calling this method will clear previous calls.

The amount of embeds must not exceed EMBED_COUNT_LIMIT. The total character length of each embed must not exceed EMBED_TOTAL_LENGTH characters. Additionally, the internal fields also have character limits. Refer to Discord Docs/Embed Limits for more information.

Errors

Returns an error of type TooManyEmbeds if there are too many embeds.

Otherwise, refer to the errors section of twilight_validate::embed::embed for a list of errors that may occur.

pub fn flags(self, flags: MessageFlags) -> Self

Set the message’s flags.

The only supported flags are SUPPRESS_EMBEDS and SUPPRESS_NOTIFICATIONS.

pub fn payload_json(self, payload_json: &'a [u8]) -> Self

JSON encoded body of any additional request fields.

If this method is called, all other fields are ignored, except for attachments. See Discord Docs/Uploading Files.

Examples

See ExecuteWebhook::payload_json for examples.

pub fn sticker_ids(self, sticker_ids: &'a [Id<StickerMarker>]) -> Self

Set the IDs of up to 3 guild stickers.

Errors

Returns an error of type StickersInvalid if the length is invalid.