Struct twilight_http::request::channel::update_channel::UpdateChannel

pub struct UpdateChannel<'a> { /* fields omitted */ }

Update a channel.

All fields are optional. The minimum length of the name is 2 UTF-16 characters and the maximum is 100 UTF-16 characters.

Errors

Returns a UpdateChannelError::NameInvalid when the length of the name is either fewer than 2 UTF-16 characters or more than 100 UTF-16 characters.

Returns a UpdateChannelError::RateLimitPerUserInvalid when the seconds of the rate limit per user is more than 21600.

Returns a UpdateChannelError::TopicInvalid when the length of the topic is more than 1024 UTF-16 characters.

Implementations

impl<'a> UpdateChannel<'a>

pub fn bitrate(self, bitrate: u64) -> Self

Set the bitrate of the channel. Applicable to voice channels only.

pub fn name(self, name: impl Into<String>) -> Result<Self, UpdateChannelError>

Set the name.

The minimum length is 2 UTF-16 characters and the maximum is 100 UTF-16 characters.

Errors

Returns UpdateChannelError::NameInvalid if the name length is too short or too long.

pub fn nsfw(self, nsfw: bool) -> Self

Set whether the channel is marked as NSFW.

pub fn parent_id(self, parent_id: impl Into<Option<ChannelId>>) -> Self

If this is specified, and the parent ID is a ChannelType::CategoryChannel, move this channel to a child of the category channel.

pub fn permission_overwrites(
self,
permission_overwrites: Vec<PermissionOverwrite>
) -> Self

Set the permission overwrites of a channel. This will overwrite all permissions that the channel currently has, so use with caution!

pub fn position(self, position: u64) -> Self

Set the position of the channel.

Positions are numerical and zero-indexed. If you place a channel at position 2, channels 2-n will shift down one position and the initial channel will take its place.

pub fn rate_limit_per_user(
self,
rate_limit_per_user: u64
) -> Result<Self, UpdateChannelError>

Set the number of seconds that a user must wait before before they are able to send another message.

The minimum is 0 and the maximum is 21600. Refer to the discord docs for more details. This is also known as “Slow Mode”.

Errors

Returns UpdateChannelError::RateLimitPerUserInvalid if the amount is greater than 21600.

pub fn topic(self, topic: impl Into<String>) -> Result<Self, UpdateChannelError>

Set the topic.

The maximum length is 1024 UTF-16 characters. Refer to the discord docs for more details.

Errors

Returns UpdateChannelError::TopicInvalid if the topic length is too long.

pub fn user_limit(self, user_limit: u64) -> Self

For voice channels, set the user limit.

Set to 0 for no limit. Limit can otherwise be between 1 and 99 inclusive. Refer to the discord docs for more details.

pub fn video_quality_mode(self, video_quality_mode: VideoQualityMode) -> Self

Set the VideoQualityMode for the voice channel.

pub fn kind(self, kind: ChannelType) -> Self

Set the kind of channel.

Only conversion between ChannelType::GuildText and ChannelType::GuildNews is possible, and only if the guild has the NEWS feature enabled. Refer to the discord docs for more details.