Module twilight_http::request::guild::integration

Structs

CreateGuildIntegration
Create a guild integration from the current user to the guild.

DeleteGuildIntegration
Delete an integration for a guild, by the integration's id.

GetGuildIntegrations
Get the guild's integrations.

SyncGuildIntegration
Synchronize a guild's integration by its id.

UpdateGuildIntegration
Update a guild's integration, by its id.