Module twilight_http::request::guild::member

Re-exports

pub use self::get_guild_members::GetGuildMembers;

pub use self::update_guild_member::UpdateGuildMember;

Modules

get_guild_members

update_guild_member

Structs

AddRoleToMember
Add a role to a member in a guild.

GetMember
Get a member of a guild, by id.

RemoveMember
Kick a member from a guild, by their id.

RemoveRoleFromMember
Remove a role from a member in a guild, by id.