Struct twilight_http::request::guild::CreateGuildChannel

pub struct CreateGuildChannel<'a> { /* private fields */ }

Create a new request to create a guild channel.

All fields are optional except for name. The minimum length of the name is 1 UTF-16 characters and the maximum is 100 UTF-16 characters.

Implementations

impl<'a> CreateGuildChannel<'a>

pub fn available_tags(self, available_tags: &'a [ForumTag]) -> Self

Set the available tags for the forum.

pub fn bitrate(self, bitrate: u32) -> Self

For voice and stage channels, set the bitrate of the channel.

Must be at least 8000.

Errors

Returns an error of type BitrateInvalid if the bitrate is invalid.

pub fn default_auto_archive_duration( self, auto_archive_duration: AutoArchiveDuration, ) -> Self

Set the default auto archive duration for newly created threads in the channel.

Automatic archive durations are not locked behind the guild’s boost level.

pub fn default_forum_layout(self, default_forum_layout: ForumLayout) -> Self

Set the default forum layout for new forum channels.

pub fn default_reaction_emoji( self, default_reaction_emoji: &'a DefaultReaction, ) -> Self

Set the default reaction emoji for new forum threads.

pub fn default_sort_order(self, default_sort_order: ForumSortOrder) -> Self

Set the default sort order for newly created forum channels.

pub fn default_thread_rate_limit_per_user( self, default_thread_rate_limit_per_user: u16, ) -> Self

Set the default number of seconds that a user must wait before before they are able to send another message in any newly-created thread in the channel.

This field is only applicable for text, announcement, media, and forum channels. The minimum is 0 and the maximum is 21600. This is also known as “Slow Mode”. See Discord Docs/Channel Object.

pub fn kind(self, kind: ChannelType) -> Self

Set the kind of channel.

pub fn nsfw(self, nsfw: bool) -> Self

Set whether the channel is marked as NSFW.

pub fn parent_id(self, parent_id: Id<ChannelMarker>) -> Self

If this is specified, and the parent ID is a ChannelType::CategoryChannel, create this channel as a child of the category channel.

pub fn permission_overwrites( self, permission_overwrites: &'a [PermissionOverwrite], ) -> Self

Set the permission overwrites of a channel.

pub fn position(self, position: u64) -> Self

Set the position of the channel.

Positions are numerical and zero-indexed. If you place a channel at position 2, channels 2-n will shift down one position and the initial channel will take its place.

pub fn rate_limit_per_user(self, rate_limit_per_user: u16) -> Self

Set the number of seconds that a user must wait before before they are able to send another message.

The minimum is 0 and the maximum is 21600. This is also known as “Slow Mode”. See Discord Docs/Channel Object.

Errors

Returns an error of type RateLimitPerUserInvalid if the name is invalid.

pub fn rtc_region(self, rtc_region: &'a str) -> Self

For voice and stage channels, set the channel’s RTC region.

pub fn topic(self, topic: &'a str) -> Self

Set the topic.

The maximum length is 1024 UTF-16 characters. See Discord Docs/Channel Object.

Errors

Returns an error of type TopicInvalid if the name is invalid.

pub fn user_limit(self, user_limit: u16) -> Self

For voice channels, set the user limit.

Set to 0 for no limit. Limit can otherwise be between 1 and 99 inclusive. See Discord Docs/Modify Channel for more details.

pub fn video_quality_mode(self, video_quality_mode: VideoQualityMode) -> Self

For voice channels, set the channel’s video quality mode.