Module twilight_http::request::prelude

Re-exports

pub use super::channel::invite::*;

pub use super::channel::message::*;

pub use super::channel::reaction::*;

pub use super::channel::webhook::*;

pub use super::channel::*;

pub use super::guild::ban::*;

pub use super::guild::emoji::*;

pub use super::guild::integration::*;

pub use super::guild::member::*;

pub use super::guild::role::*;

pub use super::guild::*;

pub use super::user::*;

Structs

GetGateway
Get information about the gateway, optionally with additional information detailing the number of shards to use and sessions remaining.

GetGatewayAuthed
Get information about the gateway, authenticated as a bot user.

GetVoiceRegions
Get a list of voice regions that can be used when creating a guild.