Struct twilight_http::request::scheduled_event::CreateGuildStageInstanceScheduledEvent

pub struct CreateGuildStageInstanceScheduledEvent<'a>(/* private fields */);

Create a stage instance scheduled event in a guild.

Implementations

impl<'a> CreateGuildStageInstanceScheduledEvent<'a>

pub fn description(self, description: &'a str) -> Self

Set the description of the event.

Must be between 1 and 1000 characters in length.

Errors

Returns an error of type ScheduledEventDescription if the description is invalid.

pub fn image(self, image: &'a str) -> Self

Set the cover image of the event.

This must be a Data URI, in the form of data:image/{type};base64,{data} where {type} is the image MIME type and {data} is the base64-encoded image. See Discord Docs/Image Data.

pub fn scheduled_end_time(self, scheduled_end_time: &'a Timestamp) -> Self

Set the scheduled end time of the event.

This is not a required field for stage instance events.