Struct twilight_http::request::sticker::get_nitro_sticker_packs::GetNitroStickerPacks

pub struct GetNitroStickerPacks<'a> { /* private fields */ }

Returns a list of sticker packs available to Nitro subscribers.

Examples

use twilight_http::Client;

let client = Client::new("my token".to_owned());

let packs = client.nitro_sticker_packs().await?.model().await?;

println!("{}", packs.sticker_packs.len());

Implementations

impl<'a> GetNitroStickerPacks<'a>

pub fn exec(self) -> ResponseFuture<StickerPackListing>

๐Deprecated since 0.14.0: use .await or into_future instead

Execute the request, returning a future resolving to a Response.