Module twilight_http::request::template

Structs

CreateGuildFromTemplate

Create a new guild based on a template.

CreateTemplate

Create a template from the current state of the guild.

DeleteTemplate

Delete a template by ID and code.

GetTemplate

Get a template by its code.

GetTemplates

Get a list of templates in a guild, by ID.

SyncTemplate

Sync a template to the current state of the guild, by ID and code.

UpdateTemplate

Update the template’s metadata, by ID and code.