Module twilight_http::response::marker

Markers denoting the type of response body.

Markers are used depending on the type of response that Twilight expects from an endpoint. For example, DeleteRole responses have no body, so an EmptyBody marker is used in the Response.

Structs

EmptyBody

Marker that a response has no body. Responses with this marker can’t be deserialized.

ListBody

Marker that a response has a list of something.