Struct twilight_http::response::Response

pub struct Response<T> { /* fields omitted */ }

Response wrapper containing helper functions over the HTTP client’s response.

This exists so that it can be determined whether to deserialize the body. This is useful when you don’t need the body and therefore don’t want to spend the time to deserialize it.

Examples

use std::env;
use twilight_http::Client;

let client = Client::new(env::var("DISCORD_TOKEN")?);
let response = client.user(user_id).exec().await?;
println!("status code: {}", response.status());

let user = response.model().await?;
println!("username: {}#{:04}", user.name, user.discriminator);

Implementations

impl<T> Response<T>

pub fn headers(&self) -> HeaderIter<'_>

Iterator of the response headers.

pub fn status(&self) -> StatusCode

Status code of the response.

pub fn bytes(self) -> BytesFuture

Consume the response and accumulate the chunked body into bytes.

For a textual representation of the response body text should be preferred.

Examples

Count the number of bytes in a response body:

use std::env;
use twilight_http::Client;

let client = Client::new(env::var("DISCORD_TOKEN")?);
let response = client.user(user_id).exec().await?;
let bytes = response.bytes().await?;

println!("size of body: {}", bytes.len());

Errors

Returns a DeserializeBodyErrorType::Chunking error type if the response body could not be entirely read.

pub fn text(self) -> TextFuture

Consume the response and accumulate the body into a string.

For the raw bytes of the response body bytes should be preferred.

Examples

Print the textual response from getting the current user:

use std::env;
use twilight_http::Client;

let client = Client::new(env::var("DISCORD_TOKEN")?);
let response = client.current_user().exec().await?;
let text = response.text().await?;

println!("body: {}", text);

Errors

Returns a DeserializeBodyErrorType::BodyNotUtf8 error type if the response body is not UTF-8 valid.

Returns a DeserializeBodyErrorType::Chunking error type if the response body could not be entirely read.

impl<T: DeserializeOwned> Response<T>

pub fn model(self) -> ModelFuture<T>

Consume the response, chunking the body and then deserializing it into the request’s matching model.

Errors

Returns a DeserializeBodyErrorType::Chunking error type if the response body could not be entirely read.

Returns a DeserializeBodyErrorType::Deserializing error type if the response body could not be deserialized into the target model.

impl<T: DeserializeOwned> Response<ListBody<T>>

pub fn models(self) -> ModelFuture<Vec<T>>

Consume the response, chunking the body and then deserializing it into a list of something.

Errors

Returns a DeserializeBodyErrorType::Chunking error type if the response body could not be entirely read.

Returns a DeserializeBodyErrorType::Deserializing error type if the response body could not be deserialized into a list of something.

impl Response<MemberBody>

pub fn model(self) -> MemberFuture

Consume the response, chunking the body and then deserializing it into a member.

Errors

Returns a DeserializeBodyErrorType::Chunking error type if the response body could not be entirely read.

Returns a DeserializeBodyErrorType::Deserializing error type if the response body could not be deserialized into a member.

impl Response<MemberListBody>

pub fn models(self) -> MemberListFuture

Consume the response, chunking the body and then deserializing it into a list of members.

Errors

Returns a DeserializeBodyErrorType::Chunking error type if the response body could not be entirely read.

Returns a DeserializeBodyErrorType::Deserializing error type if the response body could not be deserialized into a list of members.