Struct twilight_http::client::InteractionClient

pub struct InteractionClient<'a> { /* private fields */ }

Client interface for using interactions.

Examples

Retrieve the application ID and then use an interaction request:

use std::env;
use twilight_http::Client;
use twilight_model::id::Id;

let client = Client::new(env::var("DISCORD_TOKEN")?);
let application_id = Id::new(123);

let interaction_client = client.interaction(application_id);

let commands = interaction_client
    .get_global_commands()
    .exec()
    .await?
    .models()
    .await?;

println!("there are {} global commands", commands.len());

Implementations

impl<'a> InteractionClient<'a>

pub const fn interaction_callback(
&'a self,
interaction_id: Id<InteractionMarker>,
interaction_token: &'a str,
response: &'a InteractionResponse
) -> InteractionCallback<'a>

Respond to an interaction, by ID and token.

For variants of InteractionResponse that contain a CallbackData, there is an associated builder in the twilight-util crate.

pub const fn get_interaction_original(
&'a self,
interaction_token: &'a str
) -> GetOriginalResponse<'a>

Get the original message, by its token.

pub const fn update_interaction_original(
&'a self,
interaction_token: &'a str
) -> UpdateOriginalResponse<'a>

Edit the original message, by its token.

pub const fn followup_message(
&'a self,
interaction_token: &'a str,
message_id: Id<MessageMarker>
) -> GetFollowupMessage<'a>

Get a followup message of an interaction.

pub const fn delete_interaction_original(
&'a self,
interaction_token: &'a str
) -> DeleteOriginalResponse<'a>

Delete the original message, by its token.

pub const fn create_followup_message(
&'a self,
interaction_token: &'a str
) -> CreateFollowupMessage<'a>

Create a followup message, by an interaction token.

pub const fn update_followup_message(
&'a self,
interaction_token: &'a str,
message_id: Id<MessageMarker>
) -> UpdateFollowupMessage<'a>

Edit a followup message, by an interaction token.

pub const fn delete_followup_message(
&'a self,
interaction_token: &'a str,
message_id: Id<MessageMarker>
) -> DeleteFollowupMessage<'a>

Delete a followup message by interaction token and the message’s ID.

pub const fn create_guild_command(
&'a self,
guild_id: Id<GuildMarker>
) -> CreateGuildCommand<'a>

Create a new command in a guild.

pub const fn get_guild_command(
&self,
guild_id: Id<GuildMarker>,
command_id: Id<CommandMarker>
) -> GetGuildCommand<'_>

Fetch a guild command for your application.

pub const fn get_guild_commands(
&self,
guild_id: Id<GuildMarker>
) -> GetGuildCommands<'_>

Fetch all commands for a guild, by ID.

pub const fn update_guild_command(
&self,
guild_id: Id<GuildMarker>,
command_id: Id<CommandMarker>
) -> UpdateGuildCommand<'_>

Edit a command in a guild, by ID.

You must specify a name and description. See Discord Docs/Edit Guild Application Command.

pub const fn delete_guild_command(
&self,
guild_id: Id<GuildMarker>,
command_id: Id<CommandMarker>
) -> DeleteGuildCommand<'_>

Delete a command in a guild, by ID.

pub const fn set_guild_commands(
&'a self,
guild_id: Id<GuildMarker>,
commands: &'a [Command]
) -> SetGuildCommands<'a>

Set a guild’s commands.

This method is idempotent: it can be used on every start, without being ratelimited if there aren’t changes to the commands.

The Command struct has an associated builder in the twilight-util crate.

pub const fn create_global_command(&'a self) -> CreateGlobalCommand<'a>

Create a new global command.

pub const fn get_global_command(
&self,
command_id: Id<CommandMarker>
) -> GetGlobalCommand<'_>

Fetch a global command for your application.

pub const fn get_global_commands(&self) -> GetGlobalCommands<'_>

Fetch all global commands for your application.

pub const fn update_global_command(
&self,
command_id: Id<CommandMarker>
) -> UpdateGlobalCommand<'_>

Edit a global command, by ID.

You must specify a name and description. See Discord Docs/Edit Global Application Command.

pub const fn delete_global_command(
&self,
command_id: Id<CommandMarker>
) -> DeleteGlobalCommand<'_>

Delete a global command, by ID.

pub const fn set_global_commands(
&'a self,
commands: &'a [Command]
) -> SetGlobalCommands<'a>

Set global commands.

This method is idempotent: it can be used on every start, without being ratelimited if there aren’t changes to the commands.

The Command struct has an associated builder in the twilight-util crate.

pub const fn get_command_permissions(
&self,
guild_id: Id<GuildMarker>,
command_id: Id<CommandMarker>
) -> GetCommandPermissions<'_>

Fetch command permissions for a command from the current application in a guild.

pub const fn get_guild_command_permissions(
&self,
guild_id: Id<GuildMarker>
) -> GetGuildCommandPermissions<'_>

Fetch command permissions for all commands from the current application in a guild.

pub fn update_command_permissions(
&'a self,
guild_id: Id<GuildMarker>,
command_id: Id<CommandMarker>,
permissions: &'a [CommandPermissions]
) -> Result<UpdateCommandPermissions<'a>, CommandValidationError>

Update command permissions for a single command in a guild.

This overwrites the command permissions so the full set of permissions have to be sent every time.

pub fn set_command_permissions(
&'a self,
guild_id: Id<GuildMarker>,
permissions: &'a [(Id<CommandMarker>, CommandPermissions)]
) -> Result<SetCommandPermissions<'a>, CommandValidationError>

Update command permissions for all commands in a guild.

This overwrites the command permissions so the full set of permissions have to be sent every time.

Returns an error of type CountInvalid if too many commands have been provided. The maximum amount is defined by GUILD_COMMAND_LIMIT.