Enum twilight_http::error::ErrorType

#[non_exhaustive]
pub enum ErrorType {

    BuildingRequest,
    ChunkingResponse,
    CreatingHeader {
        name: String,
    },
    Formatting,
    Json,
    Parsing {
        body: Vec<u8>,
    },
    Ratelimiting,
    RequestCanceled,
    RequestError,
    RequestTimedOut,
    Response {
        body: Vec<u8>,
        error: ApiError,
        status: StatusCode,
    },
    ServiceUnavailable {
        response: Response<Body>,
    },
    Unauthorized,

}

Type of Error that occurred.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

BuildingRequest

ChunkingResponse

CreatingHeader

Fields of CreatingHeader

name: String

Formatting

Json

Parsing

Fields of Parsing

body: Vec<u8>

Ratelimiting

RequestCanceled

RequestError

RequestTimedOut

Response

Fields of Response

body: Vec<u8>error: ApiErrorstatus: StatusCode

ServiceUnavailable

API service is unavailable. Consider re-sending the request at a later time.

This may occur during Discord API stability incidents.

Fields of ServiceUnavailable

response: Response<Body>

Unauthorized

Token in use has become revoked or is otherwise invalid.

This can occur if a bot token is invalidated or an access token expires or is revoked. Recreate the client to configure a new token.