Module twilight_http::request

Modules

channel

guild

prelude

user

Structs

GetGateway
Get information about the gateway, optionally with additional information detailing the number of shards to use and sessions remaining.

GetGatewayAuthed
Get information about the gateway, authenticated as a bot user.

GetUserApplicationInfo

GetVoiceRegions
Get a list of voice regions that can be used when creating a guild.

Request

Enums

AuditLogReasonError
The error created when a reason can not be used as configured.

Traits

AuditLogReason