Module twilight_http::request::application::command

Re-exports

pub use self::create_global_command::CreateGlobalCommand;

pub use self::create_guild_command::CreateGuildCommand;

Modules

create_global_command

create_guild_command

Structs

DeleteGlobalCommand

Delete a global command, by ID.

DeleteGuildCommand

Delete a command in a guild, by ID.

GetCommandPermissions

Fetch command permissions for a command from the current application in a guild.

GetGlobalCommand

Retrieve a global command for an application.

GetGlobalCommands

Retrieve all global commands for an application.

GetGuildCommand

Retrieve a global command for an application.

GetGuildCommandPermissions

Get command permissions for all commands from the current application in a guild.

GetGuildCommands

Fetch all commands for a guild, by ID.

SetGlobalCommands

Set global commands.

SetGuildCommands

Set a guild’s commands.

UpdateCommandPermissions

Update command permissions for a single command in a guild.

UpdateGlobalCommand

Edit a global command, by ID.

UpdateGuildCommand

Edit a command in a guild, by ID.