Module twilight_http::request::channel

Re-exports

pub use self::update_channel::UpdateChannel;

Modules

invite

message

reaction

stage

update_channel

webhook

Structs

CreatePin

Create a new pin in a channel.

CreateTypingTrigger

Fire a Typing Start event in the channel.

DeleteChannel

Delete a channel by ID.

DeleteChannelPermission

Clear the permissions for a target ID in a channel.

DeleteChannelPermissionConfigured

Clear the permissions for a target ID in a channel.

DeletePin

Delete a pin in a channel, by ID.

FollowNewsChannel

Follow a news channel by ChannelIds.

GetChannel

Get a channel by its ID.

GetPins

Get the pins of a channel.

UpdateChannelPermission

Update the permissions for a role or a user in a channel.

UpdateChannelPermissionConfigured

Created when either member or role is called on a DeleteChannelPermission struct.