Struct twilight_http::request::channel::message::create_message::CreateMessage

pub struct CreateMessage<'a> { /* private fields */ }

Send a message to a channel.

The message must include at least one of attachments, content, components, embeds, or sticker_ids.

Example

use twilight_http::Client;
use twilight_model::id::Id;

let client = Client::new("my token".to_owned());

let channel_id = Id::new(123);
let message = client
    .create_message(channel_id)
    .content("Twilight is best pony")?
    .tts(true)
    .await?;

Implementations

impl<'a> CreateMessage<'a>

pub const fn allowed_mentions( self, allowed_mentions: Option<&'a AllowedMentions> ) -> Self

Specify the [AllowedMentions] for the message.

Unless otherwise called, the request will use the client’s default allowed mentions. Set to None to ignore this default.

pub fn attachments( self, attachments: &'a [Attachment] ) -> Result<Self, MessageValidationError>

Attach multiple files to the message.

Calling this method will clear previous calls.

Errors

Returns an error of type AttachmentDescriptionTooLarge if the attachments’s description is too large.

Returns an error of type AttachmentFilename if any filename is invalid.

pub fn components( self, components: &'a [Component] ) -> Result<Self, MessageValidationError>

Set the message’s list of [Component]s.

Calling this method will clear previous calls.

Errors

Refer to the errors section of [twilight_validate::component::component] for a list of errors that may be returned as a result of validating each provided component.

pub fn content(self, content: &'a str) -> Result<Self, MessageValidationError>

Set the message’s content.

The maximum length is 2000 UTF-16 characters.

Errors

Returns an error of type ContentInvalid if the content length is too long.

pub fn embeds(self, embeds: &'a [Embed]) -> Result<Self, MessageValidationError>

Set the message’s list of embeds.

Calling this method will clear previous calls.

The amount of embeds must not exceed EMBED_COUNT_LIMIT. The total character length of each embed must not exceed EMBED_TOTAL_LENGTH characters. Additionally, the internal fields also have character limits. See Discord Docs/Embed Limits.

Errors

Returns an error of type TooManyEmbeds if there are too many embeds.

Otherwise, refer to the errors section of [twilight_validate::embed::embed] for a list of errors that may occur.

pub const fn fail_if_not_exists(self, fail_if_not_exists: bool) -> Self

Whether to fail sending if the reply no longer exists.

Defaults to true.

pub const fn flags(self, flags: MessageFlags) -> Self

Set the message’s flags.

The only supported flags are SUPPRESS_EMBEDS and SUPPRESS_NOTIFICATIONS.

pub const fn nonce(self, nonce: u64) -> Self

Attach a nonce to the message, for optimistic message sending.

pub const fn payload_json(self, payload_json: &'a [u8]) -> Self

JSON encoded body of any additional request fields.

If this method is called, all other fields are ignored, except for attachments. See Discord Docs/Uploading Files.

Examples

See ExecuteWebhook::payload_json for examples.

pub const fn reply(self, other: Id<MessageMarker>) -> Self

Specify the ID of another message to create a reply to.

pub fn sticker_ids( self, sticker_ids: &'a [Id<StickerMarker>] ) -> Result<Self, MessageValidationError>

Set the IDs of up to 3 guild stickers.

Errors

Returns an error of type StickersInvalid if the length is invalid.

pub const fn tts(self, tts: bool) -> Self

Specify true if the message is TTS.

pub fn exec(self) -> ResponseFuture<Message>

👎Deprecated since 0.14.0: use .await or into_future instead

Execute the request, returning a future resolving to a Response.