Module twilight_http::request::channel::reaction

Structs

CreateReaction

Create a reaction in a [Id<ChannelMarker>] on a [Id<MessageMarker>].

DeleteAllReaction

Remove all reactions of a specified emoji from a message.

DeleteAllReactions

Delete all reactions by all users on a message.

DeleteReaction

Delete one reaction by a user on a message.

GetReactions

Get a list of users that reacted to a message with an emoji.

Enums

RequestReactionType

Handle a reaction of either a custom or unicode emoji.