Module twilight_http::request::channel::reaction

Re-exports

pub use self::get_reactions::GetReactions;

Modules

get_reactions

Structs

CreateReaction
Create a reaction in a ChannelId on a MessageId.

DeleteAllReaction
Remove all reactions of a specified emoji from a message.

DeleteAllReactions
Delete all reactions by all users on a message.

DeleteReaction
Delete one reaction by a user on a message.

Enums

RequestReactionType