Module twilight_http::request::channel::webhook

Re-exports

pub use self::execute_webhook::ExecuteWebhook;

pub use self::update_webhook_message::UpdateWebhookMessage;

Modules

execute_webhook

update_webhook_message

Update a message created by a webhook via execution.

Structs

CreateWebhook

Create a webhook in a channel.

DeleteWebhook

Delete a webhook by its ID.

DeleteWebhookMessage

Delete a message created by a webhook.

ExecuteWebhookAndWait

Execute a webhook, sending a message to its channel, and then wait for the message to be created.

GetChannelWebhooks

Get all the webhooks of a channel.

GetWebhook

Get a webhook by ID.

GetWebhookMessage

Get a webhook message by webhook ID, token, and message ID.

UpdateWebhook

Update a webhook by ID.

UpdateWebhookWithToken

Update a webhook, with a token, by ID.