Module twilight_http::request::guild

Re-exports

pub use self::create_guild::CreateGuild;

pub use self::update_guild_channel_positions::UpdateGuildChannelPositions;

Modules

auto_moderation

ban

create_guild

emoji

integration

member

role

sticker

update_guild_channel_positions

user

Structs

CreateGuildChannel

Create a new request to create a guild channel.

CreateGuildPrune

Begin a guild prune.

DeleteGuild

Delete a guild permanently. The user must be the owner.

GetActiveThreads

Returns all active threads in the guild.

GetAuditLog

Get the audit log for a guild.

GetGuild

Get information about a guild.

GetGuildChannels

Get the channels in a guild.

GetGuildInvites

Get information about the invites of a guild.

GetGuildPreview

For public guilds, get the guild preview.

GetGuildPruneCount

Get the counts of guild members to be pruned.

GetGuildVanityUrl

Get a guild’s vanity url, if there is one.

GetGuildVoiceRegions

Get voice region data for the guild.

GetGuildWebhooks

Get the webhooks of a guild.

GetGuildWelcomeScreen

Get the guild’s welcome screen.

GetGuildWidget

Get a guild’s widget

GetGuildWidgetSettings

Get a guild’s widget settings.

UpdateCurrentMember

Update the user’s member in a guild.

UpdateGuild

Update a guild.

UpdateGuildMfa

Update a guild’s MFA level.

UpdateGuildWelcomeScreen

Update the guild’s welcome screen.

UpdateGuildWidgetSettings

Modify a guild’s widget settings.