Module twilight_http::request::guild::create_guild

Structs

CategoryFields

Category channel fields sent to Discord.

CategoryFieldsBuilder

A builder for a category channel, and its children.

CategoryFieldsError

Error creating category fields.

CreateGuild

Create a new request to create a guild.

CreateGuildError

The error returned when the guild can not be created as configured.

GuildChannelFieldsBuilder

A builder for a list of channels.

RoleFields

Role fields sent to Discord.

RoleFieldsBuilder

A builder for role fields.

RoleFieldsError

Error building role fields.

TextFields

Text channel fields sent to Discord.

TextFieldsBuilder

A builder for text fields.

TextFieldsError

Error building text fields.

VoiceFields

Voice channel fields sent to Discord.

VoiceFieldsBuilder

A builder for voice fields.

VoiceFieldsError

Error building voice fields.

Enums

CategoryFieldsErrorType

Type of CategoryFieldsError that occurred.

CreateGuildErrorType

Type of CreateGuildError that occurred.

GuildChannelFields

Variants of channel fields sent to Discord.

RoleFieldsErrorType

Type of RoleFieldsError that occurred.

TextFieldsErrorType

Type of TextFieldsError that occurred.

VoiceFieldsErrorType

Type of VoiceFieldsError that occurred.