Struct twilight_http::request::guild::CreateGuildChannel

pub struct CreateGuildChannel<'a> { /* private fields */ }

Create a new request to create a guild channel.

All fields are optional except for name. The minimum length of the name is 1 UTF-16 characters and the maximum is 100 UTF-16 characters.

Implementations

impl<'a> CreateGuildChannel<'a>

pub const fn available_tags(self, available_tags: &'a [ForumTag]) -> Self

Set the available tags for the forum.

pub const fn bitrate(self, bitrate: u32) -> Result<Self, ChannelValidationError>

For voice and stage channels, set the bitrate of the channel.

Must be at least 8000.

Errors

Returns an error of type BitrateInvalid if the bitrate is invalid.

pub const fn default_auto_archive_duration(
self,
auto_archive_duration: AutoArchiveDuration
) -> Self

Set the default auto archive duration for newly created threads in the channel.

Automatic archive durations are not locked behind the guild’s boost level.

pub const fn default_reaction_emoji(
self,
default_reaction_emoji: &'a DefaultReaction
) -> Self

Set the default reaction emoji for new forum threads.

pub const fn kind(self, kind: ChannelType) -> Self

Set the kind of channel.

pub const fn nsfw(self, nsfw: bool) -> Self

Set whether the channel is marked as NSFW.

pub const fn parent_id(self, parent_id: Id<ChannelMarker>) -> Self

If this is specified, and the parent ID is a ChannelType::CategoryChannel, create this channel as a child of the category channel.

pub const fn permission_overwrites(
self,
permission_overwrites: &'a [PermissionOverwrite]
) -> Self

Set the permission overwrites of a channel.

pub const fn position(self, position: u64) -> Self

Set the position of the channel.

Positions are numerical and zero-indexed. If you place a channel at position 2, channels 2-n will shift down one position and the initial channel will take its place.

pub const fn rate_limit_per_user(
self,
rate_limit_per_user: u16
) -> Result<Self, ChannelValidationError>

Set the number of seconds that a user must wait before before they are able to send another message.

The minimum is 0 and the maximum is 21600. This is also known as “Slow Mode”. See Discord Docs/Channel Object.

Errors

Returns an error of type RateLimitPerUserInvalid if the name is invalid.

pub const fn rtc_region(self, rtc_region: &'a str) -> Self

For voice and stage channels, set the channel’s RTC region.

pub fn topic(self, topic: &'a str) -> Result<Self, ChannelValidationError>

Set the topic.

The maximum length is 1024 UTF-16 characters. See Discord Docs/Channel Object.

Errors

Returns an error of type TopicInvalid if the name is invalid.

pub const fn user_limit(self, user_limit: u16) -> Self

For voice channels, set the user limit.

Set to 0 for no limit. Limit can otherwise be between 1 and 99 inclusive. See Discord Docs/Modify Channel for more details.

pub const fn video_quality_mode(
self,
video_quality_mode: VideoQualityMode
) -> Self

For voice channels, set the channel’s video quality mode.

pub fn exec(self) -> ResponseFuture<Channel>

Execute the request, returning a future resolving to a Response.