Struct twilight_http::request::guild::UpdateGuild

pub struct UpdateGuild<'a> { /* private fields */ }

Update a guild.

All endpoints are optional. See Discord Docs/Modify Guild.

Implementations

impl<'a> UpdateGuild<'a>

pub const fn afk_channel_id(
self,
afk_channel_id: Option<Id<ChannelMarker>>
) -> Self

Set the voice channel where AFK voice users are sent.

pub const fn afk_timeout(self, afk_timeout: u64) -> Self

Set how much time it takes for a voice user to be considered AFK.

pub const fn banner(self, banner: Option<&'a str>) -> Self

Set the banner.

This is a base64 encoded 16:9 PNG or JPEG image. Pass None to remove the banner.

The server must have the BANNER feature.

pub const fn default_message_notifications(
self,
default_message_notifications: Option<DefaultMessageNotificationLevel>
) -> Self

Set the default message notification level. See Discord Docs/Create Guild for more information.

pub const fn discovery_splash(self, discovery_splash: Option<&'a str>) -> Self

Set the guild’s discovery splash image.

Requires the guild to have the DISCOVERABLE feature enabled.

pub const fn explicit_content_filter(
self,
explicit_content_filter: Option<ExplicitContentFilter>
) -> Self

Set the explicit content filter level.

pub const fn features(self, features: &'a [&'a str]) -> Self

Set the enabled features of the guild.

Attempting to add or remove the GuildFeature::Community feature requires the Permissions::ADMINISTRATOR permission.

Attempting to add or remove the GuildFeature::Discoverable feature requires the Permissions::ADMINISTRATOR permission. Additionally the guild must pass all the discovery requirements.

Attempting to add or remove the GuildFeature::InvitesDisabled feature requires the Permissions::MANAGE_GUILD permission.

pub const fn icon(self, icon: Option<&'a str>) -> Self

Set the icon.

This must be a Data URI, in the form of data:image/{type};base64,{data} where {type} is the image MIME type and {data} is the base64-encoded image. See Discord Docs/Image Data.

pub fn name(self, name: &'a str) -> Result<Self, ValidationError>

Set the name of the guild.

The minimum length is 1 UTF-16 character and the maximum is 100 UTF-16 characters.

Errors

Returns an error of type GuildName if the name length is too short or too long.

pub const fn owner_id(self, owner_id: Id<UserMarker>) -> Self

Transfer ownership to another user.

Only works if the current user is the owner.

pub const fn splash(self, splash: Option<&'a str>) -> Self

Set the guild’s splash image.

Requires the guild to have the INVITE_SPLASH feature enabled.

pub const fn system_channel(
self,
system_channel_id: Option<Id<ChannelMarker>>
) -> Self

Set the channel where events such as welcome messages are posted.

pub const fn system_channel_flags(
self,
system_channel_flags: Option<SystemChannelFlags>
) -> Self

Set the guild’s [SystemChannelFlags].

pub const fn rules_channel(
self,
rules_channel_id: Option<Id<ChannelMarker>>
) -> Self

Set the rules channel.

Requires the guild to be PUBLIC. See Discord Docs/Modify Guild.

pub const fn public_updates_channel(
self,
public_updates_channel_id: Option<Id<ChannelMarker>>
) -> Self

Set the public updates channel.

Requires the guild to be PUBLIC.

pub const fn preferred_locale(self, preferred_locale: Option<&'a str>) -> Self

Set the preferred locale for the guild.

Defaults to en-US. Requires the guild to be PUBLIC.

pub const fn verification_level(
self,
verification_level: Option<VerificationLevel>
) -> Self

Set the verification level.

See Discord Docs/Guild Object.

pub const fn premium_progress_bar_enabled(
self,
premium_progress_bar_enabled: bool
) -> Self

Set whether the premium progress bar is enabled.

pub fn exec(self) -> ResponseFuture<PartialGuild>

👎Deprecated since 0.14.0: use .await or into_future instead

Execute the request, returning a future resolving to a Response.