Struct twilight_http::request::Request

pub struct Request { /* private fields */ }

Implementations

impl Request

pub fn builder(route: &Route<'_>) -> RequestBuilder

Create a new request builder.

Examples

Create a request to create a message with a content of “test” in a channel with an ID of 1:

use twilight_http::{request::Request, routing::Route};

let body = br#"{
    "content": "test"
}"#
.to_vec();

let request = Request::builder(&Route::CreateMessage { channel_id: 1 })
    .body(body)
    .build();

pub fn from_route(route: &Route<'_>) -> Self

Create a request from only its route information.

If you need to set additional configurations like the body then use builder.

Examples

Create a request to get a message with an ID of 2 in a channel with an ID of 1:

use twilight_http::{request::Request, routing::Route};

let request = Request::from_route(&Route::GetMessage {
    channel_id: 1,
    message_id: 2,
});

pub fn body(&self) -> Option<&[u8]>

Body of the request, if any.

pub const fn form(&self) -> Option<&Form>

Multipart form of the request, if any.

pub const fn headers(&self) -> Option<&HeaderMap<HeaderValue>>

Headers to set in the request, if any.

pub const fn method(&self) -> Method

Method when sending the request.

pub fn path(&self) -> &str

String path of the full URL.

pub const fn ratelimit_path(&self) -> &Path

Path used for ratelimiting.

pub const fn use_authorization_token(&self) -> bool

Whether to use the client’s authorization token in the request.