Struct twilight_http::request::RequestBuilder

pub struct RequestBuilder(_);

Builder to create a customized request.

Examples

Create a request to create a message with a content of “test” in a channel with an ID of 1:

use twilight_http::{request::Request, routing::Route};

let body = br#"{
    "content": "test"
}"#.to_vec();

let request = Request::builder(Route::CreateMessage {
    channel_id: 1,
}).body(body).build();

Implementations

impl RequestBuilder

#[must_use = "request has not been fully built"]
pub fn new(route: Route) -> Self

Create a new request builder.

#[must_use = "request information is not useful on its own and must be acted on"]
pub fn build(self) -> Request

Consume the builder, returning the built request.

#[must_use = "request has not been fully built"]
pub fn body(self, body: Vec<u8>) -> Self

Set the contents of the body.

#[must_use = "request has not been fully built"]
pub fn form(self, form: Form) -> Self

Set the multipart form.

#[must_use = "request has not been fully built"]
pub fn headers(
self,
iter: impl Iterator<Item = (HeaderName, HeaderValue)>
) -> Self

Set the headers to add.

#[must_use = "request has not been fully built"]
pub fn json(self, to: &impl Serialize) -> Result<Self>

Set the body, to be serialized as JSON.

Errors

Returns an ErrorType::Json error type if the value could not be serialized as JSON.

pub fn use_authorization_token(self, use_authorization_token: bool) -> Self

Whether to use the client’s authorization token in the request, if one is set.

This is primarily useful for executing webhooks.