Enum twilight_http::routing::Path

pub enum Path {
    ChannelsId(u64),
    ChannelsIdInvites(u64),
    ChannelsIdMessages(u64),
    ChannelsIdMessagesBulkDelete(u64),
    ChannelsIdMessagesId(Method, u64),
    ChannelsIdMessagesIdReactions(u64),
    ChannelsIdMessagesIdReactionsUserIdType(u64),
    ChannelsIdPermissionsOverwriteId(u64),
    ChannelsIdPins(u64),
    ChannelsIdPinsMessageId(u64),
    ChannelsIdTyping(u64),
    ChannelsIdWebhooks(u64),
    Gateway,
    GatewayBot,
    Guilds,
    GuildsId(u64),
    GuildsIdBans(u64),
    GuildsIdBansId(u64),
    GuildsIdAuditLogs(u64),
    GuildsIdBansUserId(u64),
    GuildsIdChannels(u64),
    GuildsIdWidget(u64),
    GuildsIdEmojis(u64),
    GuildsIdEmojisId(u64),
    GuildsIdIntegrations(u64),
    GuildsIdIntegrationsId(u64),
    GuildsIdIntegrationsIdSync(u64),
    GuildsIdInvites(u64),
    GuildsIdMembers(u64),
    GuildsIdMembersId(u64),
    GuildsIdMembersIdRolesId(u64),
    GuildsIdMembersMeNick(u64),
    GuildsIdPreview(u64),
    GuildsIdPrune(u64),
    GuildsIdRegions(u64),
    GuildsIdRoles(u64),
    GuildsIdRolesId(u64),
    GuildsIdVanityUrl(u64),
    GuildsIdWebhooks(u64),
    InvitesCode,
    UsersId,
    UsersIdConnections,
    UsersIdChannels,
    UsersIdGuilds,
    UsersIdGuildsId,
    VoiceRegions,
    WebhooksId(u64),
}

An enum representing a path, most useful for ratelimiting implementations.

Variants

ChannelsId(u64)

Operating on a channel.

ChannelsIdInvites(u64)

Operating on a channel's invites.

ChannelsIdMessages(u64)

Operating on a channel's messages.

ChannelsIdMessagesBulkDelete(u64)

Operating on a channel's messages by bulk deleting.

ChannelsIdMessagesId(Method, u64)

Operating on an individual channel's message.

ChannelsIdMessagesIdReactions(u64)

Operating on an individual channel's message's reactions.

ChannelsIdMessagesIdReactionsUserIdType(u64)

Operating on an individual channel's message's reactions while specifying the user ID and emoji type.

ChannelsIdPermissionsOverwriteId(u64)

Operating on a channel's permission overwrites by ID.

ChannelsIdPins(u64)

Operating on a channel's pins.

ChannelsIdPinsMessageId(u64)

Operating on a channel's individual pinned message.

ChannelsIdTyping(u64)

Operating on a channel's typing indicator.

ChannelsIdWebhooks(u64)

Operating on a channel's webhooks.

Gateway

Operating with the gateway information.

GatewayBot

Operating with the gateway information tailored to the current user.

Guilds

Operating on the guild resource.

GuildsId(u64)

Operating on one of user's guilds.

GuildsIdBans(u64)

GuildsIdBansId(u64)

GuildsIdAuditLogs(u64)

GuildsIdBansUserId(u64)

GuildsIdChannels(u64)

GuildsIdWidget(u64)

GuildsIdEmojis(u64)

GuildsIdEmojisId(u64)

GuildsIdIntegrations(u64)

GuildsIdIntegrationsId(u64)

GuildsIdIntegrationsIdSync(u64)

GuildsIdInvites(u64)

GuildsIdMembers(u64)

GuildsIdMembersId(u64)

GuildsIdMembersIdRolesId(u64)

GuildsIdMembersMeNick(u64)

GuildsIdPreview(u64)

GuildsIdPrune(u64)

GuildsIdRegions(u64)

GuildsIdRoles(u64)

GuildsIdRolesId(u64)

GuildsIdVanityUrl(u64)

GuildsIdWebhooks(u64)

InvitesCode

UsersId

UsersIdConnections

UsersIdChannels

UsersIdGuilds

Operating on the state of a guild that the user is in.

UsersIdGuildsId

Operating on the state of a guild that the user is in.

VoiceRegions

Operating on the voice regions available to the current user.

WebhooksId(u64)

Operating on a webhook.