Enum twilight_http::routing::Route

#[non_exhaustive]
pub enum Route<'a> {

    AddGuildMember {
        guild_id: u64,
        user_id: u64,
    },
    AddMemberRole {
        guild_id: u64,
        role_id: u64,
        user_id: u64,
    },
    AddThreadMember {
        channel_id: u64,
        user_id: u64,
    },
    CreateBan {
        delete_message_days: Option<u16>,
        guild_id: u64,
        user_id: u64,
    },
    CreateChannel {
        guild_id: u64,
    },
    CreateEmoji {
        guild_id: u64,
    },
    CreateGlobalCommand {
        application_id: u64,
    },
    CreateGuild,
    CreateGuildCommand {
        application_id: u64,
        guild_id: u64,
    },
    CreateGuildFromTemplate {
        template_code: &'a str,
    },
    CreateGuildIntegration {
        guild_id: u64,
    },
    CreateGuildPrune {
        compute_prune_count: Option<bool>,
        days: Option<u16>,
        guild_id: u64,
        include_roles: &'a [Id<RoleMarker>],
    },
    CreateGuildScheduledEvent {
        guild_id: u64,
    },
    CreateGuildSticker {
        guild_id: u64,
    },
    CreateInvite {
        channel_id: u64,
    },
    CreateMessage {
        channel_id: u64,
    },
    CreatePrivateChannel,
    CreateThread {
        channel_id: u64,
    },
    CreateThreadFromMessage {
        channel_id: u64,
        message_id: u64,
    },
    CreateReaction {
        channel_id: u64,
        emoji: &'a RequestReactionType<'a>,
        message_id: u64,
    },
    CreateRole {
        guild_id: u64,
    },
    CreateStageInstance,
    CreateTemplate {
        guild_id: u64,
    },
    CreateTypingTrigger {
        channel_id: u64,
    },
    CreateWebhook {
        channel_id: u64,
    },
    CrosspostMessage {
        channel_id: u64,
        message_id: u64,
    },
    DeleteBan {
        guild_id: u64,
        user_id: u64,
    },
    DeleteChannel {
        channel_id: u64,
    },
    DeleteEmoji {
        emoji_id: u64,
        guild_id: u64,
    },
    DeleteGlobalCommand {
        application_id: u64,
        command_id: u64,
    },
    DeleteGuild {
        guild_id: u64,
    },
    DeleteGuildCommand {
        application_id: u64,
        command_id: u64,
        guild_id: u64,
    },
    DeleteGuildIntegration {
        guild_id: u64,
        integration_id: u64,
    },
    DeleteGuildScheduledEvent {
        guild_id: u64,
        scheduled_event_id: u64,
    },
    DeleteGuildSticker {
        guild_id: u64,
        sticker_id: u64,
    },
    DeleteInvite {
        code: &'a str,
    },
    DeleteInteractionOriginal {
        application_id: u64,
        interaction_token: &'a str,
    },
    DeleteMessage {
        channel_id: u64,
        message_id: u64,
    },
    DeleteMessages {
        channel_id: u64,
    },
    DeleteMessageReactions {
        channel_id: u64,
        message_id: u64,
    },
    DeleteMessageSpecificReaction {
        channel_id: u64,
        emoji: &'a RequestReactionType<'a>,
        message_id: u64,
    },
    DeletePermissionOverwrite {
        channel_id: u64,
        target_id: u64,
    },
    DeleteReactionCurrentUser {
        channel_id: u64,
        emoji: &'a RequestReactionType<'a>,
        message_id: u64,
    },
    DeleteReaction {
        channel_id: u64,
        emoji: &'a RequestReactionType<'a>,
        message_id: u64,
        user_id: u64,
    },
    DeleteRole {
        guild_id: u64,
        role_id: u64,
    },
    DeleteStageInstance {
        channel_id: u64,
    },
    DeleteTemplate {
        guild_id: u64,
        template_code: &'a str,
    },
    DeleteWebhookMessage {
        message_id: u64,
        thread_id: Option<u64>,
        token: &'a str,
        webhook_id: u64,
    },
    DeleteWebhook {
        token: Option<&'a str>,
        webhook_id: u64,
    },
    ExecuteWebhook {
        thread_id: Option<u64>,
        token: &'a str,
        wait: Option<bool>,
        webhook_id: u64,
    },
    FollowNewsChannel {
        channel_id: u64,
    },
    GetActiveThreads {
        guild_id: u64,
    },
    GetAuditLogs {
        action_type: Option<u64>,
        before: Option<u64>,
        guild_id: u64,
        limit: Option<u16>,
        user_id: Option<u64>,
    },
    GetBan {
        guild_id: u64,
        user_id: u64,
    },
    GetBans {
        guild_id: u64,
    },
    GetBansWithParameters {
        after: Option<u64>,
        before: Option<u64>,
        limit: Option<u16>,
        guild_id: u64,
    },
    GetChannel {
        channel_id: u64,
    },
    GetChannelInvites {
        channel_id: u64,
    },
    GetChannelWebhooks {
        channel_id: u64,
    },
    GetChannels {
        guild_id: u64,
    },
    GetCommandPermissions {
        application_id: u64,
        command_id: u64,
        guild_id: u64,
    },
    GetCurrentUserApplicationInfo,
    GetCurrentUser,
    GetCurrentUserGuildMember {
        guild_id: u64,
    },
    GetEmoji {
        emoji_id: u64,
        guild_id: u64,
    },
    GetEmojis {
        guild_id: u64,
    },
    GetFollowupMessage {
        application_id: u64,
        interaction_token: &'a str,
        thread_id: Option<u64>,
        message_id: u64,
    },
    GetGateway,
    GetGatewayBot,
    GetGlobalCommand {
        application_id: u64,
        command_id: u64,
    },
    GetGlobalCommands {
        application_id: u64,
        with_localizations: Option<bool>,
    },
    GetGuild {
        guild_id: u64,
        with_counts: bool,
    },
    GetGuildCommand {
        application_id: u64,
        command_id: u64,
        guild_id: u64,
    },
    GetGuildCommandPermissions {
        application_id: u64,
        guild_id: u64,
    },
    GetGuildCommands {
        application_id: u64,
        guild_id: u64,
        with_localizations: Option<bool>,
    },
    GetGuildWidget {
        guild_id: u64,
    },
    GetGuildIntegrations {
        guild_id: u64,
    },
    GetGuildInvites {
        guild_id: u64,
    },
    GetGuildMembers {
        after: Option<u64>,
        guild_id: u64,
        limit: Option<u16>,
        presences: Option<bool>,
    },
    GetGuildPreview {
        guild_id: u64,
    },
    GetGuildPruneCount {
        days: Option<u16>,
        guild_id: u64,
        include_roles: &'a [Id<RoleMarker>],
    },
    GetGuildRoles {
        guild_id: u64,
    },
    GetGuildScheduledEvent {
        guild_id: u64,
        scheduled_event_id: u64,
        with_user_count: bool,
    },
    GetGuildScheduledEvents {
        guild_id: u64,
        with_user_count: bool,
    },
    GetGuildScheduledEventUsers {
        after: Option<u64>,
        before: Option<u64>,
        guild_id: u64,
        limit: Option<u16>,
        scheduled_event_id: u64,
        with_member: bool,
    },
    GetGuildSticker {
        guild_id: u64,
        sticker_id: u64,
    },
    GetGuildStickers {
        guild_id: u64,
    },
    GetGuildVanityUrl {
        guild_id: u64,
    },
    GetGuildVoiceRegions {
        guild_id: u64,
    },
    GetGuildWelcomeScreen {
        guild_id: u64,
    },
    GetGuildWebhooks {
        guild_id: u64,
    },
    GetGuilds {
        after: Option<u64>,
        before: Option<u64>,
        limit: Option<u16>,
    },
    GetInteractionOriginal {
        application_id: u64,
        interaction_token: &'a str,
    },
    GetInvite {
        code: &'a str,
        with_counts: bool,
    },
    GetInviteWithExpiration {
        code: &'a str,
        with_counts: bool,
        with_expiration: bool,
    },
    GetMember {
        guild_id: u64,
        user_id: u64,
    },
    GetMessage {
        channel_id: u64,
        message_id: u64,
    },
    GetMessages {
        after: Option<u64>,
        around: Option<u64>,
        before: Option<u64>,
        channel_id: u64,
        limit: Option<u16>,
    },
    GetNitroStickerPacks,
    GetPins {
        channel_id: u64,
    },
    GetJoinedPrivateArchivedThreads {
        before: Option<u64>,
        channel_id: u64,
        limit: Option<u64>,
    },
    GetPrivateArchivedThreads {
        before: Option<&'a str>,
        channel_id: u64,
        limit: Option<u64>,
    },
    GetPublicArchivedThreads {
        before: Option<&'a str>,
        channel_id: u64,
        limit: Option<u64>,
    },
    GetReactionUsers {
        after: Option<u64>,
        channel_id: u64,
        emoji: &'a RequestReactionType<'a>,
        limit: Option<u16>,
        message_id: u64,
    },
    GetStageInstance {
        channel_id: u64,
    },
    GetSticker {
        sticker_id: u64,
    },
    GetTemplate {
        template_code: &'a str,
    },
    GetTemplates {
        guild_id: u64,
    },
    GetThreadMember {
        channel_id: u64,
        user_id: u64,
    },
    GetThreadMembers {
        channel_id: u64,
    },
    GetUser {
        user_id: u64,
    },
    GetUserConnections,
    GetUserPrivateChannels,
    GetVoiceRegions,
    GetWebhook {
        token: Option<&'a str>,
        webhook_id: u64,
    },
    GetWebhookMessage {
        message_id: u64,
        thread_id: Option<u64>,
        token: &'a str,
        webhook_id: u64,
    },
    InteractionCallback {
        interaction_id: u64,
        interaction_token: &'a str,
    },
    JoinThread {
        channel_id: u64,
    },
    LeaveGuild {
        guild_id: u64,
    },
    LeaveThread {
        channel_id: u64,
    },
    PinMessage {
        channel_id: u64,
        message_id: u64,
    },
    RemoveMember {
        guild_id: u64,
        user_id: u64,
    },
    RemoveMemberRole {
        guild_id: u64,
        role_id: u64,
        user_id: u64,
    },
    RemoveThreadMember {
        channel_id: u64,
        user_id: u64,
    },
    SearchGuildMembers {
        guild_id: u64,
        limit: Option<u16>,
        query: &'a str,
    },
    SetGlobalCommands {
        application_id: u64,
    },
    SetGuildCommands {
        application_id: u64,
        guild_id: u64,
    },
    SyncGuildIntegration {
        guild_id: u64,
        integration_id: u64,
    },
    SyncTemplate {
        guild_id: u64,
        template_code: &'a str,
    },
    UnpinMessage {
        channel_id: u64,
        message_id: u64,
    },
    UpdateChannel {
        channel_id: u64,
    },
    UpdateCommandPermissions {
        application_id: u64,
        command_id: u64,
        guild_id: u64,
    },
    UpdateCurrentMember {
        guild_id: u64,
    },
    UpdateCurrentUser,
    UpdateCurrentUserVoiceState {
        guild_id: u64,
    },
    UpdateEmoji {
        emoji_id: u64,
        guild_id: u64,
    },
    UpdateGlobalCommand {
        application_id: u64,
        command_id: u64,
    },
    UpdateGuild {
        guild_id: u64,
    },
    UpdateGuildChannels {
        guild_id: u64,
    },
    UpdateGuildCommand {
        application_id: u64,
        command_id: u64,
        guild_id: u64,
    },
    UpdateGuildWidget {
        guild_id: u64,
    },
    UpdateGuildIntegration {
        guild_id: u64,
        integration_id: u64,
    },
    UpdateGuildScheduledEvent {
        guild_id: u64,
        scheduled_event_id: u64,
    },
    UpdateGuildSticker {
        guild_id: u64,
        sticker_id: u64,
    },
    UpdateGuildWelcomeScreen {
        guild_id: u64,
    },
    UpdateInteractionOriginal {
        application_id: u64,
        interaction_token: &'a str,
    },
    UpdateMember {
        guild_id: u64,
        user_id: u64,
    },
    UpdateMessage {
        channel_id: u64,
        message_id: u64,
    },
    UpdateNickname {
        guild_id: u64,
    },
    UpdatePermissionOverwrite {
        channel_id: u64,
        target_id: u64,
    },
    UpdateRole {
        guild_id: u64,
        role_id: u64,
    },
    UpdateRolePositions {
        guild_id: u64,
    },
    UpdateStageInstance {
        channel_id: u64,
    },
    UpdateTemplate {
        guild_id: u64,
        template_code: &'a str,
    },
    UpdateUserVoiceState {
        guild_id: u64,
        user_id: u64,
    },
    UpdateWebhookMessage {
        message_id: u64,
        thread_id: Option<u64>,
        token: &'a str,
        webhook_id: u64,
    },
    UpdateWebhook {
        token: Option<&'a str>,
        webhook_id: u64,
    },

}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AddGuildMember

Fields

guild_id: u64

user_id: u64

Route information to add a user to a guild.

AddMemberRole

Fields

guild_id: u64

The ID of the guild.

role_id: u64

The ID of the role.

user_id: u64

The ID of the user.

Route information to add a role to guild member.

AddThreadMember

Fields

channel_id: u64

ID of the thread.

user_id: u64

ID of the member.

Route information to add a member to a thread.

CreateBan

Fields

delete_message_days: Option<u16>

The number of days’ worth of the user’s messages to delete in the guild’s channels.

guild_id: u64

The ID of the guild.

user_id: u64

The ID of the user.

Route information to create a ban on a user in a guild.

CreateChannel

Fields

guild_id: u64

The ID of the guild.

Route information to create a channel in a guild.

CreateEmoji

Fields

guild_id: u64

The ID of the guild.

Route information to create an emoji in a guild.

CreateGlobalCommand

Fields

application_id: u64

The ID of the owner application.

Route information to create a global command.

CreateGuild

Route information to create a guild.

CreateGuildCommand

Fields

application_id: u64

The ID of the owner application.

guild_id: u64

The ID of the guild.

Route information to create a guild command.

CreateGuildFromTemplate

Fields

template_code: &'a str

Code of the template.

Route information to create a guild from a template.

CreateGuildIntegration

Fields

guild_id: u64

The ID of the guild.

Route information to create a guild’s integration.

CreateGuildPrune

Fields

compute_prune_count: Option<bool>

Whether to compute the number of pruned users.

days: Option<u16>

The number of days that a user must be offline before being able to be pruned.

guild_id: u64

The ID of the guild.

include_roles: &'a [Id<RoleMarker>]

The roles to filter the prune by.

A user must have at least one of these roles to be able to be pruned.

Route information to create a prune in a guild.

CreateGuildScheduledEvent

Fields

guild_id: u64

ID of the guild.

Route information to create a scheduled event in a guild.

CreateGuildSticker

Fields

guild_id: u64

ID of the guild.

Route information to create a sticker in a guild.

CreateInvite

Fields

channel_id: u64

The ID of the channel.

Route information to create an invite to a channel.

CreateMessage

Fields

channel_id: u64

The ID of the channel.

Route information to create a message in a channel.

CreatePrivateChannel

Route information to create a private channel.

CreateThread

Fields

channel_id: u64

ID of the channel.

Route information to create a thread in a channel.

CreateThreadFromMessage

Fields

channel_id: u64

ID of the channel.

message_id: u64

ID of the message.

Route information to create a thread from a message.

CreateReaction

Fields

channel_id: u64

The ID of the channel.

emoji: &'a RequestReactionType<'a>

The URI encoded custom or unicode emoji.

message_id: u64

The ID of the message.

Route information to create a reaction on a message.

CreateRole

Fields

guild_id: u64

The ID of the guild.

Route information to create a role in a guild.

CreateStageInstance

Route information to create a stage instance.

CreateTemplate

Fields

guild_id: u64

The ID of the guild.

Route information to create a guild template.

CreateTypingTrigger

Fields

channel_id: u64

The ID of the channel.

Route information to create a typing trigger in a channel.

CreateWebhook

Fields

channel_id: u64

The ID of the channel.

Route information to create a webhook in a channel.

CrosspostMessage

Fields

channel_id: u64

The ID of the channel.

message_id: u64

The ID of the message.

Route information to crosspost a message to following guilds.

DeleteBan

Fields

guild_id: u64

The ID of the guild.

user_id: u64

The ID of the user.

Route information to delete a ban on a user in a guild.

DeleteChannel

Fields

channel_id: u64

The ID of the channel.

Route information to delete a channel.

DeleteEmoji

Fields

emoji_id: u64

The ID of the emoji.

guild_id: u64

The ID of the guild.

Route information to delete a guild’s custom emoji.

DeleteGlobalCommand

Fields

application_id: u64

The ID of the owner application.

command_id: u64

The ID of the command.

Route information to delete a global command.

DeleteGuild

Fields

guild_id: u64

The ID of the guild.

Route information to delete a guild.

DeleteGuildCommand

Fields

application_id: u64

The ID of the owner application.

command_id: u64

The ID of the command.

guild_id: u64

The ID of the guild.

Route information to delete a guild command.

DeleteGuildIntegration

Fields

guild_id: u64

The ID of the guild.

integration_id: u64

The ID of the integration.

Route information to delete a guild integration.

DeleteGuildScheduledEvent

Fields

guild_id: u64

ID of the guild.

scheduled_event_id: u64

ID of the scheduled event.

Route information to delete a scheduled event in a guild.

DeleteGuildSticker

Fields

guild_id: u64

ID of the guild.

sticker_id: u64

ID of the sticker.

Route information to delete a guild sticker.

DeleteInvite

Fields

code: &'a str

The unique invite code.

Route information to delete an invite.

DeleteInteractionOriginal

Fields

application_id: u64

The ID of the owner application

interaction_token: &'a str

The token of the interaction.

Route information to delete the original interaction response.

DeleteMessage

Fields

channel_id: u64

The ID of the channel.

message_id: u64

The ID of the message.

Route information to delete a channel’s message.

DeleteMessages

Fields

channel_id: u64

The ID of the channel.

Route information to bulk delete messages in a channel.

DeleteMessageReactions

Fields

channel_id: u64

The ID of the channel.

message_id: u64

The ID of the message.

Route information to delete all of the reactions on a message.

DeleteMessageSpecificReaction

Fields

channel_id: u64

The ID of the channel.

emoji: &'a RequestReactionType<'a>

The URI encoded custom or unicode emoji.

message_id: u64

The ID of the message.

Route information to delete all of the reactions on a message with a specific emoji.

DeletePermissionOverwrite

Fields

channel_id: u64

The ID of the channel.

target_id: u64

The ID of the target role or user.

Route information to delete a permission overwrite for a role or user in a channel.

DeleteReactionCurrentUser

Fields

channel_id: u64

ID of the channel.

emoji: &'a RequestReactionType<'a>

URI encoded custom or unicode emoji.

message_id: u64

ID of the message.

Route information to delete the current user’s reaction on a message.

DeleteReaction

Fields

channel_id: u64

The ID of the channel.

emoji: &'a RequestReactionType<'a>

The URI encoded custom or unicode emoji.

message_id: u64

The ID of the message.

user_id: u64

The ID of the user.

Route information to delete a user’s reaction on a message.

DeleteRole

Fields

guild_id: u64

The ID of the guild.

role_id: u64

The ID of the role.

Route information to delete a guild’s role.

DeleteStageInstance

Fields

channel_id: u64

ID of the stage channel.

Route information to delete a stage instance.

DeleteTemplate

Fields

guild_id: u64

The ID of the guild.

template_code: &'a str

The target template code.

Route information to delete a guild template.

DeleteWebhookMessage

Fields

message_id: u64

thread_id: Option<u64>

ID of the thread channel, if there is one.

token: &'a str

webhook_id: u64

Route information to delete a message created by a webhook.

DeleteWebhook

Fields

token: Option<&'a str>

The token of the webhook.

webhook_id: u64

The ID of the webhook.

Route information to delete a webhook.

ExecuteWebhook

Fields

thread_id: Option<u64>

ID of the thread channel, if there is one.

token: &'a str

The token of the webhook.

wait: Option<bool>

Whether to wait for a message response.

webhook_id: u64

The ID of the webhook.

Route information to execute a webhook by ID and token.

FollowNewsChannel

Fields

channel_id: u64

The ID of the channel to follow.

Route information to follow a news channel.

GetActiveThreads

Fields

guild_id: u64

ID of the guild.

Route information to get active threads in a channel.

GetAuditLogs

Fields

action_type: Option<u64>

The type of action to get audit logs for.

before: Option<u64>

The maximum ID of audit logs to get.

guild_id: u64

The ID of the guild.

limit: Option<u16>

The maximum number of audit logs to get.

user_id: Option<u64>

The ID of the user, if specified.

Route information to get a paginated list of audit logs in a guild.

GetBan

Fields

guild_id: u64

The ID of the guild.

user_id: u64

The ID of the user.

Route information to get information about a single ban in a guild.

GetBans

Fields

guild_id: u64

The ID of the guild.

Route information to get a guild’s bans.

GetBansWithParameters

Fields

after: Option<u64>

User ID after which to retrieve bans.

before: Option<u64>

User ID before which to retrieve bans.

limit: Option<u16>

Maximum number of bans to retrieve.

guild_id: u64

ID of the guild.

Route information to get a guild’s bans with parameters.

GetChannel

Fields

channel_id: u64

The ID of the channel.

Route information to get a channel.

GetChannelInvites

Fields

channel_id: u64

The ID of the channel.

Route information to get a channel’s invites.

GetChannelWebhooks

Fields

channel_id: u64

The ID of the channel.

Route information to get a channel’s webhooks.

GetChannels

Fields

guild_id: u64

The ID of the guild.

Route information to get a guild’s channels.

GetCommandPermissions

Fields

application_id: u64

The ID of the application.

command_id: u64

The ID of the command.

guild_id: u64

The ID of the guild.

Route information to get permissions of a specific guild command.

GetCurrentUserApplicationInfo

Route information to get info about application the current bot user belongs to

GetCurrentUser

Route information to get the current user.

GetCurrentUserGuildMember

Fields

guild_id: u64

ID of the guild.

Route information to get the current user as a member object within a guild.

GetEmoji

Fields

emoji_id: u64

The ID of the emoji.

guild_id: u64

The ID of the guild.

Route information to get an emoji by ID within a guild.

GetEmojis

Fields

guild_id: u64

The ID of the guild.

Route information to get a guild’s emojis.

GetFollowupMessage

Fields

application_id: u64

ID of the application.

interaction_token: &'a str

Token of the interaction.

thread_id: Option<u64>

ID of the thread channel, if there is one.

message_id: u64

ID of the followup message.

Route to get a followup message for an interaction.

GetGateway

Route information to get basic gateway information.

GetGatewayBot

Route information to get gateway information tailored to the current user.

GetGlobalCommand

Fields

application_id: u64

ID of the owner application.

command_id: u64

ID of the command.

Route information to get a global command for an application.

GetGlobalCommands

Fields

application_id: u64

The ID of the owner application.

with_localizations: Option<bool>

Whether to include full localization dictionaries.

GetGuild

Fields

guild_id: u64

The ID of the guild.

with_counts: bool

Whether to include approximate member and presence counts for the guild.

Route information to get a guild.

GetGuildCommand

Fields

application_id: u64

ID of the owner application.

command_id: u64

ID of the command.

guild_id: u64

ID of the guild.

Route information to get a guild command.

GetGuildCommandPermissions

Fields

application_id: u64

The ID of the application.

guild_id: u64

The ID of the guild.

Route information to get permissions of all guild commands.

GetGuildCommands

Fields

application_id: u64

The ID of the owner application.

guild_id: u64

The ID of the guild.

with_localizations: Option<bool>

Whether to include full localization dictionaries.

Route information to get guild commands.

GetGuildWidget

Fields

guild_id: u64

The ID of the guild.

Route information to get a guild’s widget.

GetGuildIntegrations

Fields

guild_id: u64

The ID of the guild.

Route information to get a guild’s integrations.

GetGuildInvites

Fields

guild_id: u64

The ID of the guild.

Route information to get a guild’s invites.

GetGuildMembers

Fields

after: Option<u64>

The minimum ID of members to get.

guild_id: u64

The ID of the guild.

limit: Option<u16>

The maximum number of members to get.

presences: Option<bool>

Whether to get the members’ presences.

Route information to get a guild’s members.

GetGuildPreview

Fields

guild_id: u64

The ID of the guild.

Route information to get a guild’s preview.

GetGuildPruneCount

Fields

days: Option<u16>

The number of days that a user must be offline before being able to be pruned.

guild_id: u64

The ID of the guild.

include_roles: &'a [Id<RoleMarker>]

The roles to filter the prune by.

A user must have at least one of these roles to be able to be pruned.

Route information to get the number of members that would be pruned, filtering by inactivity and users with one of the provided roles.

GetGuildRoles

Fields

guild_id: u64

The ID of the guild.

Route information to get a guild’s roles.

GetGuildScheduledEvent

Fields

guild_id: u64

ID of the guild.

scheduled_event_id: u64

with_user_count: bool

Whether to include user counts.

Route information to get a guild scheduled event.

GetGuildScheduledEvents

Fields

guild_id: u64

ID of the guild.

with_user_count: bool

Whether to include user counts.

Route information to get a guild’s scheduled events.

GetGuildScheduledEventUsers

Fields

after: Option<u64>

Get members after this ID.

before: Option<u64>

Get members before this ID.

guild_id: u64

ID of the guild.

limit: Option<u16>

Maximum amount of members to get.

scheduled_event_id: u64

ID of the scheduled event.

with_member: bool

Whether to return a member object.

Route information to get a guild scheduled event’s members.

GetGuildSticker

Fields

guild_id: u64

ID of the guild.

sticker_id: u64

ID of the stickers.

Route information to get a guild’s sticker.

GetGuildStickers

Fields

guild_id: u64

ID of the guild.

Route information to get a guild’s stickers.

GetGuildVanityUrl

Fields

guild_id: u64

The ID of the guild.

Route information to get a guild’s vanity URL.

GetGuildVoiceRegions

Fields

guild_id: u64

The ID of the guild.

Route information to get a guild’s available voice regions.

GetGuildWelcomeScreen

Fields

guild_id: u64

ID of the guild.

Route information to get a guild’s welcome screen.

GetGuildWebhooks

Fields

guild_id: u64

The ID of the guild.

Route information to get a guild’s webhooks.

GetGuilds

Fields

after: Option<u64>

The minimum ID of guilds to get.

before: Option<u64>

The maximum ID of guilds to get.

limit: Option<u16>

The maximum number of guilds to get.

Route information to get a paginated list of guilds.

GetInteractionOriginal

Fields

application_id: u64

ID of the owner application.

interaction_token: &'a str

Token of the interaction.

Route information to get an original interaction response message.

GetInvite

Fields

code: &'a str

The unique invite code.

with_counts: bool

Whether to retrieve statistics about the invite.

Route information to get an invite.

GetInviteWithExpiration

Fields

code: &'a str

The unique invite code.

with_counts: bool

Whether to retrieve statistics about the invite.

with_expiration: bool

Whether to retrieve the expiration date of the invite.

Route information to get an invite with an expiration.

GetMember

Fields

guild_id: u64

The ID of the guild.

user_id: u64

The ID of the user.

Route information to get a member.

GetMessage

Fields

channel_id: u64

The ID of the channel.

message_id: u64

The ID of the message.

Route information to get a single message in a channel.

GetMessages

Fields

after: Option<u64>

The minimum ID of messages to get.

around: Option<u64>

The message ID to get the messages around.

before: Option<u64>

The maximum ID of messages to get.

channel_id: u64

The ID of the channel.

limit: Option<u16>

The maximum number of messages to get.

Route information to get a paginated list of messages in a channel.

GetNitroStickerPacks

Route information to get a list of sticker packs available to Nitro subscribers.

GetPins

Fields

channel_id: u64

The ID of the channel.

Route information to get a channel’s pins.

GetJoinedPrivateArchivedThreads

Fields

before: Option<u64>

Optional timestamp to return threads before.

channel_id: u64

ID of the channel.

limit: Option<u64>

Optional maximum number of threads to return.

Route information to get joined private archived threads in a channel.

GetPrivateArchivedThreads

Fields

before: Option<&'a str>

Optional timestamp to return threads before.

channel_id: u64

ID of the channel.

limit: Option<u64>

Optional maximum number of threads to return.

Route information to get private archived threads in a channel.

GetPublicArchivedThreads

Fields

before: Option<&'a str>

Optional timestamp to return threads before.

channel_id: u64

ID of the channel.

limit: Option<u64>

Optional maximum number of threads to return.

Route information to get public archived threads in a channel.

GetReactionUsers

Fields

after: Option<u64>

The minimum ID of users to get.

channel_id: u64

The ID of the channel.

emoji: &'a RequestReactionType<'a>

The URI encoded custom or unicode emoji.

limit: Option<u16>

The maximum number of users to retrieve.

message_id: u64

The ID of the message.

Route information to get the users who reacted to a message with a specified emoji.

GetStageInstance

Fields

channel_id: u64

ID of the stage channel.

Route information to get a stage instance.

GetSticker

Fields

sticker_id: u64

ID of the sticker.

Route information to get a sticker.

GetTemplate

Fields

template_code: &'a str

The template code.

Route information to get a template.

GetTemplates

Fields

guild_id: u64

The ID of the guild.

Route information to get a list of templates from a guild.

GetThreadMember

Fields

channel_id: u64

ID of the thread.

user_id: u64

ID of the member.

Route information to get a member of a thread.

GetThreadMembers

Fields

channel_id: u64

ID of the thread.

Route information to get members of a thread.

GetUser

Fields

user_id: u64

ID of the target user.

Route information to get a user.

GetUserConnections

Route information to get the current user’s connections.

GetUserPrivateChannels

Route information to get the current user’s private channels and groups.

GetVoiceRegions

Route information to get a list of the voice regions.

GetWebhook

Fields

token: Option<&'a str>

The token of the webhook.

webhook_id: u64

The ID of the webhook.

Route information to get a webhook by ID, optionally with a token if the current user doesn’t have access to it.

GetWebhookMessage

Fields

message_id: u64

ID of the message.

thread_id: Option<u64>

ID of the thread channel, if there is one.

token: &'a str

Token of the webhook.

webhook_id: u64

ID of the webhook.

Route information to get a previously-sent webhook message.

InteractionCallback

Fields

interaction_id: u64

The ID of the interaction.

interaction_token: &'a str

The token for the interaction.

Route information to respond to an interaction.

JoinThread

Fields

channel_id: u64

ID of the thread.

Route information to join a thread as the current user.

LeaveGuild

Fields

guild_id: u64

The ID of the guild.

Route information to leave the guild.

LeaveThread

Fields

channel_id: u64

ID of the thread.

Route information to leave a thread as the current user.

PinMessage

Fields

channel_id: u64

The ID of the channel.

message_id: u64

The ID of the message.

Route information to pin a message to a channel.

RemoveMember

Fields

guild_id: u64

The ID of the guild.

user_id: u64

The ID of the user.

Route information to remove a member from a guild.

RemoveMemberRole

Fields

guild_id: u64

The ID of the guild.

role_id: u64

The ID of the role.

user_id: u64

The ID of the user.

Route information to remove a role from a member.

RemoveThreadMember

Fields

channel_id: u64

ID of the thread.

user_id: u64

ID of the member.

Route information to remove a member from a thread.

SearchGuildMembers

Fields

guild_id: u64

ID of the guild to search in.

limit: Option<u16>

Upper limit of members to query for.

query: &'a str

Query to search by.

Route information to search for members in a guild.

SetGlobalCommands

Fields

application_id: u64

The ID of the owner application.

Route information to set global commands.

SetGuildCommands

Fields

application_id: u64

The ID of the owner application.

guild_id: u64

The ID of the guild.

Route information to set guild commands.

SyncGuildIntegration

Fields

guild_id: u64

The ID of the guild.

integration_id: u64

The ID of the integration.

Route information to sync a guild’s integration.

SyncTemplate

Fields

guild_id: u64

The ID of the guild.

template_code: &'a str

The template code.

Route information to sync a template.

UnpinMessage

Fields

channel_id: u64

The ID of the channel.

message_id: u64

The ID of the message.

Route information to unpin a message from a channel.

UpdateChannel

Fields

channel_id: u64

The ID of the channel.

Route information to update a channel, such as a guild channel or group.

UpdateCommandPermissions

Fields

application_id: u64

The ID of the application.

command_id: u64

The ID of the command.

guild_id: u64

The ID of the guild.

Route information to edit permissions of a command in a guild.

UpdateCurrentMember

Fields

guild_id: u64

ID of the guild.

Route information to update the current member.

UpdateCurrentUser

Route information to update the current user.

UpdateCurrentUserVoiceState

Fields

guild_id: u64

ID of the guild.

Route information to update the current user’s voice state.

UpdateEmoji

Fields

emoji_id: u64

The ID of the emoji.

guild_id: u64

The ID of the guild.

Route information to update an emoji.

UpdateGlobalCommand

Fields

application_id: u64

The ID of the owner application.

command_id: u64

The ID of the command.

Route information to update a global command.

UpdateGuild

Fields

guild_id: u64

The ID of the guild.

Route information to update a guild.

UpdateGuildChannels

Fields

guild_id: u64

The ID of the guild.

Route information to update a guild channel.

UpdateGuildCommand

Fields

application_id: u64

The ID of the owner application.

command_id: u64

The ID of the command.

guild_id: u64

The ID of the guild.

Route information to update a guild command.

UpdateGuildWidget

Fields

guild_id: u64

The ID of the guild.

Route information to update a guild’s widget.

UpdateGuildIntegration

Fields

guild_id: u64

The ID of the guild.

integration_id: u64

The ID of the integration.

Route information to update a guild’s integration.

UpdateGuildScheduledEvent

Fields

guild_id: u64

ID of the guild.

scheduled_event_id: u64

ID of the scheduled event.

Route information to update a scheduled event in a guild.

UpdateGuildSticker

Fields

guild_id: u64

ID of the guild.

sticker_id: u64

ID of the sticker.

Route information to update a guild sticker.

UpdateGuildWelcomeScreen

Fields

guild_id: u64

ID of the guild.

Route information to update a guild’s welcome screen.

UpdateInteractionOriginal

Fields

application_id: u64

The ID of the owner application.

interaction_token: &'a str

The token for the interaction.

Update the original interaction response.

UpdateMember

Fields

guild_id: u64

The ID of the guild.

user_id: u64

The ID of the user.

Route information to update a member.

UpdateMessage

Fields

channel_id: u64

The ID of the channel.

message_id: u64

The ID of the message.

Route information to update a message.

UpdateNickname

Fields

guild_id: u64

The ID of the guild.

Route information to update the current member’s nickname.

UpdatePermissionOverwrite

Fields

channel_id: u64

The ID of the channel.

target_id: u64

The ID of the role or user.

Route information to update the permission overwrite of a role or user in a channel.

UpdateRole

Fields

guild_id: u64

The ID of the guild.

role_id: u64

The ID of the role.

Route information to update a role.

UpdateRolePositions

Fields

guild_id: u64

The ID of the guild.

Route information to update the positions of roles.

UpdateStageInstance

Fields

channel_id: u64

ID of the stage channel.

Route information to update an existing stage instance.

UpdateTemplate

Fields

guild_id: u64

The ID of the guild.

template_code: &'a str

The template code.

Route information to update a template.

UpdateUserVoiceState

Fields

guild_id: u64

ID of the guild.

user_id: u64

ID of the user.

Route information to update a user’s voice state.

UpdateWebhookMessage

Fields

message_id: u64

thread_id: Option<u64>

ID of the thread channel, if there is one.

token: &'a str

webhook_id: u64

Route information to update a message created by a webhook.

UpdateWebhook

Fields

token: Option<&'a str>

The token of the webhook.

webhook_id: u64

The ID of the webhook.

Route information to update a webhook.

Implementations

impl<'a> Route<'a>

pub const fn method(&self) -> Method

HTTP method of the route.

Examples

Assert that the GetGuild route returns Method::Get:

use twilight_http::{request::Method, routing::Route};

let route = Route::GetGuild {
    guild_id: 123,
    with_counts: false,
};

assert_eq!(Method::Get, route.method());

pub fn to_path(&self) -> Path

Typed path of the route.

Paths are used with a Ratelimiter.

Examples

Use a route’s path to retrieve a ratelimiter ticket:

use twilight_http::routing::Route;
use twilight_http_ratelimiting::{InMemoryRatelimiter, Ratelimiter};

let ratelimiter = InMemoryRatelimiter::new();
let route = Route::CreateMessage { channel_id: 123 };

// Take a ticket from the ratelimiter.
let rx = ratelimiter.ticket(route.to_path()).await?;

// Wait to be told that a request can be made...
let _tx = rx.await;

// The request can now be made.